requirements.txt
addict==2.4.0
affine==2.3.1
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.1
contourpy==1.0.6
cycler==0.11.0
dataclasses==0.6
datasets==2.7.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.29
htmlmin==0.1.12
huggingface-hub==0.11.1
idna==3.4
idna-ssl==1.1.0
imagecodecs==2023.1.23
ImageHash==4.3.1
imageio==2.22.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
ipykernel==6.19.0
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.0
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
missingno==0.5.1
mistune==2.0.4
mmcv-full==1.7.1
multidict==6.0.3
multimethod==1.9
multiprocess==0.70.14
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.6.0.66
packaging==22.0
pandas==1.5.2
pandas-profiling==3.5.0
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==4.21.11
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rasterio==1.3.4
regex==2022.10.31
requests==2.28.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.3.2.post1
stack-data==0.6.2
statsmodels==0.13.5
tangled-up-in-unicode==0.2.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.13.2
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.11.1
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
transformers==4.25.1
typeguard==2.13.3
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
visions==0.7.5
wandb==0.13.6
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.4
xxhash==3.1.0
yapf==0.32.0
yarl==1.8.2
zipp==3.11.0
